Move native stuff around
