+++ src/native/lz4/CMakeLists.txt
@@ -30,10 +30,12 @@ target_include_directories(
   "$<BUILD_INTERFACE:${LZ4_INCLUDE_DIR}>"
 )
 
-set_target_properties(
-  ${LIB_NAME}
-  PROPERTIES
-  ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}"
-)
+if(DEBUG_BUILD)
+  set_target_properties(
+    ${LIB_NAME}
+    PROPERTIES
+    ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}"
+  )
+endif()
 
 xa_add_compile_definitions(${LIB_NAME})
