+++ src/native/monodroid/monodroid-glue.cc
@@ -46,10 +46,7 @@
 #include <netinet/in.h>
 #endif
 
-//#include "java-interop-util.h"
 #include "logger.hh"
-
-//#include "monodroid.h"
 #include "util.hh"
 #include "debug.hh"
 #include "embedded-assemblies.hh"
@@ -58,36 +55,25 @@
 #include "globals.hh"
 #include "xamarin-app.hh"
 #include "timing.hh"
-//#include "xa-internal-api-impl.hh"
 #include "build-info.hh"
 #include "monovm-properties.hh"
-#include "startup-aware-lock.hh"
 #include "timing-internal.hh"
-#include "search.hh"
 #include "runtime-util.hh"
-
-//#include "xamarin_getifaddrs.h"
-
+#include "monodroid-state.hh"
+#include "pinvoke-override-api.hh"
 #include "cpp-util.hh"
 #include "strings.hh"
-
-#include "java-interop-dlfcn.h"
 
 using namespace microsoft::java_interop;
 using namespace xamarin::android;
 using namespace xamarin::android::internal;
 
-xamarin::android::mutex MonodroidRuntime::pinvoke_map_write_lock;
-
 MonoCoreRuntimeProperties MonodroidRuntime::monovm_core_properties = {
 	.trusted_platform_assemblies = nullptr,
 	.app_paths = nullptr,
 	.native_dll_search_directories = nullptr,
-	.pinvoke_override = &MonodroidRuntime::monodroid_pinvoke_override
+	.pinvoke_override = &PinvokeOverride::monodroid_pinvoke_override
 };
-
-xamarin::android::mutex MonodroidRuntime::dso_handle_write_lock;
-bool MonodroidRuntime::startup_in_progress = true;
 
 void
 MonodroidRuntime::thread_start ([[maybe_unused]] MonoProfiler *prof, [[maybe_unused]] uintptr_t tid)
@@ -1003,173 +989,6 @@ setup_gc_logging (void)
 }
 #endif
 
-force_inline unsigned int
-MonodroidRuntime::convert_dl_flags (int flags)
-{
-	unsigned int lflags = (flags & static_cast<int> (MONO_DL_LOCAL))
-		? JAVA_INTEROP_LIB_LOAD_LOCALLY
-		: JAVA_INTEROP_LIB_LOAD_GLOBALLY;
-	return lflags;
-}
-
-force_inline DSOCacheEntry*
-MonodroidRuntime::find_dso_cache_entry (hash_t hash) noexcept
-{
-	auto equal = [](DSOCacheEntry const& entry, hash_t key) -> bool { return entry.hash == key; };
-	auto less_than = [](DSOCacheEntry const& entry, hash_t key) -> bool { return entry.hash < key; };
-	ssize_t idx = Search::binary_search<DSOCacheEntry, equal, less_than> (hash, dso_cache, application_config.number_of_dso_cache_entries);
-	if (idx >= 0) {
-		return &dso_cache[idx];
-	}
-
-	return nullptr;
-}
-
-force_inline void*
-MonodroidRuntime::monodroid_dlopen_log_and_return (void *handle, char **err, const char *full_name, bool free_memory, [[maybe_unused]] bool need_api_init)
-{
-	if (handle == nullptr && err != nullptr) {
-		const char *load_error = dlerror ();
-		if (load_error == nullptr) {
-			load_error = "Unknown error";
-		}
-		*err = Util::monodroid_strdup_printf ("Could not load library '%s'. %s", full_name, load_error);
-	}
-
-	if (free_memory) {
-		delete[] full_name;
-	}
-
-	return handle;
-}
-
-force_inline void*
-MonodroidRuntime::monodroid_dlopen_ignore_component_or_load ([[maybe_unused]] hash_t name_hash, const char *name, int flags, char **err) noexcept
-{
-	if (startup_in_progress) {
-		auto ignore_component = [&](const char *label, MonoComponent component) -> bool {
-			if ((application_config.mono_components_mask & component) != component) {
-				log_info (LOG_ASSEMBLY, "Mono '%s' component requested but not packaged, ignoring", label);
-				return true;
-			}
-
-			return false;
-		};
-
-		switch (name_hash) {
-			case mono_component_debugger_hash:
-				if (ignore_component ("Debugger", MonoComponent::Debugger)) {
-					return nullptr;
-				}
-				break;
-
-			case mono_component_hot_reload_hash:
-				if (ignore_component ("Hot Reload", MonoComponent::HotReload)) {
-					return nullptr;
-				}
-				break;
-
-			case mono_component_diagnostics_tracing_hash:
-				if (ignore_component ("Diagnostics Tracing", MonoComponent::Tracing)) {
-					return nullptr;
-				}
-				break;
-		}
-	}
-
-	unsigned int dl_flags = monodroidRuntime.convert_dl_flags (flags);
-	void * handle = AndroidSystem::load_dso_from_any_directories (name, dl_flags);
-	if (handle != nullptr) {
-		return monodroid_dlopen_log_and_return (handle, err, name, false /* name_needs_free */);
-	}
-
-	handle = AndroidSystem::load_dso (name, dl_flags, false /* skip_existing_check */);
-	return monodroid_dlopen_log_and_return (handle, err, name, false /* name_needs_free */);
-}
-
-force_inline void*
-MonodroidRuntime::monodroid_dlopen (const char *name, int flags, char **err) noexcept
-{
-	hash_t name_hash = xxhash::hash (name, strlen (name));
-	log_debug (LOG_ASSEMBLY, "monodroid_dlopen: hash for name '%s' is 0x%zx", name, name_hash);
-	DSOCacheEntry *dso = find_dso_cache_entry (name_hash);
-	log_debug (LOG_ASSEMBLY, "monodroid_dlopen: hash match %sfound, DSO name is '%s'", dso == nullptr ? "not " : "", dso == nullptr ? "<unknown>" : dso->name);
-
-	if (dso == nullptr) {
-		// DSO not known at build time, try to load it
-		return monodroid_dlopen_ignore_component_or_load (name_hash, name, flags, err);
-	} else if (dso->handle != nullptr) {
-		return monodroid_dlopen_log_and_return (dso->handle, err, dso->name, false /* name_needs_free */);
-	}
-
-	if (dso->ignore) {
-		log_info (LOG_ASSEMBLY, "Request to load '%s' ignored, it is known not to exist", dso->name);
-		return nullptr;
-	}
-
-	StartupAwareLock lock (dso_handle_write_lock);
-#if defined (RELEASE)
-	if (AndroidSystem::is_embedded_dso_mode_enabled ()) {
-		DSOApkEntry *apk_entry = dso_apk_entries;
-		for (size_t i = 0; i < application_config.number_of_shared_libraries; i++) {
-			if (apk_entry->name_hash != dso->real_name_hash) {
-				apk_entry++;
-				continue;
-			}
-
-			android_dlextinfo dli;
-			dli.flags = ANDROID_DLEXT_USE_LIBRARY_FD | ANDROID_DLEXT_USE_LIBRARY_FD_OFFSET;
-			dli.library_fd = apk_entry->fd;
-			dli.library_fd_offset = apk_entry->offset;
-			dso->handle = android_dlopen_ext (dso->name, flags, &dli);
-
-			if (dso->handle != nullptr) {
-				return monodroid_dlopen_log_and_return (dso->handle, err, dso->name, false /* name_needs_free */);
-			}
-			break;
-		}
-	}
-#endif
-	unsigned int dl_flags = monodroidRuntime.convert_dl_flags (flags);
-	dso->handle = AndroidSystem::load_dso_from_any_directories (dso->name, dl_flags);
-
-	if (dso->handle != nullptr) {
-		return monodroid_dlopen_log_and_return (dso->handle, err, dso->name, false /* name_needs_free */);
-	}
-
-	dso->handle = AndroidSystem::load_dso_from_any_directories (name, dl_flags);
-	return monodroid_dlopen_log_and_return (dso->handle, err, name, false /* name_needs_free */);
-}
-
-void*
-MonodroidRuntime::monodroid_dlopen (const char *name, int flags, char **err, [[maybe_unused]] void *user_data) noexcept
-{
-	if (name == nullptr) {
-		log_warn (LOG_ASSEMBLY, "monodroid_dlopen got a null name. This is not supported in NET+");
-		return nullptr;
-	}
-
-	return monodroid_dlopen (name, flags, err);
-}
-
-void*
-MonodroidRuntime::monodroid_dlsym (void *handle, const char *name, char **err, [[maybe_unused]] void *user_data)
-{
-	void *s;
-	char *e = nullptr;
-
-	s = java_interop_lib_symbol (handle, name, &e);
-
-	if (!s && err) {
-		*err = Util::monodroid_strdup_printf ("Could not find symbol '%s': %s", name, e);
-	}
-	if (e) {
-		java_interop_free (e);
-	}
-
-	return s;
-}
-
 inline void
 MonodroidRuntime::set_environment_variable_for_directory (const char *name, jstring_wrapper &value, bool createDirectory, mode_t mode)
 {
@@ -1650,7 +1469,7 @@ MonodroidRuntime::Java_mono_android_Runtime_initInternal (JNIEnv *env, jclass kl
 	}
 
 	AndroidSystem::setup_process_args (runtimeApks);
-	mono_dl_fallback_register (monodroid_dlopen, monodroid_dlsym, nullptr, nullptr);
+	mono_dl_fallback_register (MonodroidDl::monodroid_dlopen, MonodroidDl::monodroid_dlsym, nullptr, nullptr);
 
 	set_profile_options ();
 
@@ -1731,7 +1550,7 @@ MonodroidRuntime::Java_mono_android_Runtime_initInternal (JNIEnv *env, jclass kl
 		xamarin_app_init (env, get_function_pointer_at_runtime);
 	}
 #endif // def RELEASE && def ANDROID && def NET
-	startup_in_progress = false;
+	MonodroidState::mark_startup_done ();
 }
 
 JNIEXPORT jint JNICALL
