+++ src/native/monodroid/monodroid-tracing.cc
@@ -25,9 +25,9 @@ MonodroidRuntime::log_traces (JNIEnv *env, TraceKind kind, const char *first_lin
 		std::lock_guard lock (tracing_init_lock);
 
 		char *err = nullptr;
-		void *handle = monodroid_dlopen (xamarin_native_tracing_name.data (), MONO_DL_EAGER, &err, nullptr);
+		void *handle = MonodroidDl::monodroid_dlopen (SharedConstants::xamarin_native_tracing_name.data (), MONO_DL_EAGER, &err, nullptr);
 		if (handle == nullptr) {
-			log_warn (LOG_DEFAULT, "Failed to load native tracing library '%s'. %s", xamarin_native_tracing_name, err == nullptr ? "Unknown error" : err);
+			log_warn (LOG_DEFAULT, "Failed to load native tracing library '%s'. %s", SharedConstants::xamarin_native_tracing_name, err == nullptr ? "Unknown error" : err);
 		} else {
 			load_symbol (handle, "xa_get_native_backtrace", _xa_get_native_backtrace);
 			load_symbol (handle, "xa_get_managed_backtrace", _xa_get_managed_backtrace);
