+++ src/native/pinvoke-override/dynamic.cc
@@ -1,0 +1,2 @@
+#define PINVOKE_OVERRIDE_INLINE [[gnu::noinline]]
+#include "pinvoke-override-api.hh"
