+++ src/native/runtime-base/CMakeLists.txt
@@ -6,6 +6,8 @@ set(XA_RUNTIME_BASE_SOURCES
   cpu-arch-detect.cc
   logger.cc
   shared-constants.cc
+  timing.cc
+  timing-internal.cc
   util.cc
 )
 add_clang_check_sources("${XA_RUNTIME_BASE_SOURCES}")
@@ -52,10 +54,12 @@ target_link_libraries(
   xa::xamarin-app
 )
 
-set_target_properties(
-  ${LIB_NAME}
-  PROPERTIES
-  ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}"
-)
+if(DEBUG_BUILD)
+  set_target_properties(
+    ${LIB_NAME}
+    PROPERTIES
+    ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}"
+  )
+endif()
 
 xa_add_compile_definitions(${LIB_NAME})
