+++ src/native/runtime-base/timing-internal.cc
