+++ src/native/runtime-base/timing.hh
